requirements.txt
Flask==3.0.3
pandas==2.2.2
pymongo==4.7.3
pyspark==3.5.1
python-dotenv==1.0.1
XlsxWriter==3.2.0
openpyxl==3.1.5